Enable instant search and default search on homepage
Instant search is now enabled in the Header component by setting enableInstantSearch to true. Additionally, HomePageContent will automatically perform a default search for "فنجان" when the component mounts if no search has been performed and no query is present in the URL. The implementation documentation was also updated with a detailed Arabic technical overview and recommendations.